Extract logic of shard snapshot functions into shard holder
